# Hide OpenSSL sha256 initialisation in a function-scoped `thread_local`