--- a/src/node/rpc/test/frontend_test.cpp
+++ b/src/node/rpc/test/frontend_test.cpp
@@ -5,7 +5,6 @@
 
 #define DOCTEST_CONFIG_IMPLEMENT
 #include "ccf/app_interface.h"
-#include "ccf/crypto/openssl_init.h"
 #include "ccf/ds/logger.h"
 #include "ccf/json_handler.h"
 #include "ccf/kv/map.h"
@@ -1556,18 +1555,15 @@ TEST_CASE("Manual conflicts")
   auto call_pausable = [&](
                          std::shared_ptr<ccf::SessionContext> session,
                          ccf::http_status expected_status) {
-    ccf::crypto::openssl_sha256_init();
     auto req = create_simple_request("/pausable");
     auto serialized_call = req.build_request();
     auto rpc_ctx = ccf::make_rpc_context(session, serialized_call);
     frontend.process(rpc_ctx);
     auto response = parse_response(rpc_ctx->serialise_response());
     CHECK(response.status == expected_status);
-    ccf::crypto::openssl_sha256_shutdown();
   };
 
   auto get_metrics = [&]() {
-    ccf::crypto::openssl_sha256_init();
     auto req = create_simple_request("/pausable/metrics");
     req.set_method(HTTP_GET);
     auto serialized_call = req.build_request();
@@ -1580,36 +1576,32 @@ TEST_CASE("Manual conflicts")
     ret.calls = body["calls"].get<size_t>();
     ret.retries = body["retries"].get<size_t>();
     ret.errors = body["errors"].get<size_t>();
-    ccf::crypto::openssl_sha256_shutdown();
+
     return ret;
   };
 
   auto get_value = [&](const std::string& table = TF::DST) {
-    ccf::crypto::openssl_sha256_init();
     auto tx = network.tables->create_tx();
     auto handle = tx.ro<TF::MyVals>(table);
     auto ret = handle->get(TF::KEY);
     REQUIRE(tx.commit() == ccf::kv::CommitResult::SUCCESS);
-    ccf::crypto::openssl_sha256_shutdown();
+
     return ret;
   };
 
   auto update_value =
     [&](size_t n, const std::string& table = TF::SRC, size_t key = TF::KEY) {
-      ccf::crypto::openssl_sha256_init();
       auto tx = network.tables->create_tx();
       using TF = TestManualConflictsFrontend;
       auto handle = tx.wo<TF::MyVals>(table);
       handle->put(key, n);
       REQUIRE(tx.commit() == ccf::kv::CommitResult::SUCCESS);
-      ccf::crypto::openssl_sha256_shutdown();
     };
 
   auto run_test = [&](
                     std::function<void()>&& read_write_op,
                     std::shared_ptr<ccf::SessionContext> session = user_session,
                     ccf::http_status expected_status = HTTP_STATUS_OK) {
-    ccf::crypto::openssl_sha256_init();
     frontend.registry.before_read.ready = false;
     frontend.registry.after_read.ready = false;
     frontend.registry.before_write.ready = false;
@@ -1626,7 +1618,6 @@ TEST_CASE("Manual conflicts")
     frontend.registry.after_write.wait();
 
     worker.join();
-    ccf::crypto::openssl_sha256_shutdown();
   };
 
   {
@@ -1787,11 +1778,11 @@ TEST_CASE("Manual conflicts")
 int main(int argc, char** argv)
 {
   ::threading::ThreadMessaging::init(1);
-  ccf::crypto::openssl_sha256_init();
+
   doctest::Context context;
   context.applyCommandLine(argc, argv);
   int res = context.run();
-  ccf::crypto::openssl_sha256_shutdown();
+
   if (context.shouldExit())
     return res;
   return res;