… include/ccf/crypto/openssl_init.h
… src/node/test/endorsements.cpp
… src/node/test/historical_queries.cpp
… src/node/test/history.cpp
… src/node/test/history_bench.cpp
… src/node/test/merkle_bench.cpp
… src/node/test/merkle_test.cpp
… src/node/test/receipt.cpp
… src/node/test/snapshot.cpp
… src/node/test/snapshotter.cpp
… src/pal/test/snp_attestation_validation.cpp
… src/pal/test/snp_ioctl_test.cpp
… tests/perf-system/submitter/submit.cpp