See discussion in this thread:
https://github.com/microsoft/CCF/pull/7251/files#r2322012456

I think this is much nicer, but it does seem to be slightly slower (checking a bool for `thread_local` initialisation on every call), so this is debatable.