@eddyashton 2025-09-04
## Is this actually faster?

If we only run a single digest, then the benchmarks say no - this is faster than not caching the contexts, but slower than the explicit `_init` call (since this approach introduces an implicit atomic read, to check whether the `thread_local` has been constructed).

```
## digest sha256:

 Name (* = baseline)      |   Dim   |  Total ms |  ns/op  |Baseline| Ops/second
--------------------------|--------:|----------:|--------:|-------:|----------:
 openssl_sha256_noopt *   |      32 |     0.000 |      11 |      - | 84880636.6
 openssl_sha256_opt       |      32 |     0.000 |       5 |  0.440 |192771084.3
 openssl_sha256_tl        |      32 |     0.000 |       7 |  0.599 |141592920.4
 openssl_sha256_noopt *   |     128 |     0.001 |       5 |      - |184971098.3
 openssl_sha256_opt       |     128 |     0.000 |       2 |  0.382 |484848484.8
 openssl_sha256_tl        |     128 |     0.000 |       2 |  0.403 |458781362.0
 openssl_sha256_noopt *   |     512 |     0.001 |       1 |      - |596041909.2
 openssl_sha256_opt       |     512 |     0.001 |       1 |  0.629 |948148148.1
 openssl_sha256_tl        |     512 |     0.001 |       1 |  0.698 |853333333.3
 openssl_sha256_noopt *   |    2048 |     0.002 |       1 |      - |938158497.5
 openssl_sha256_opt       |    2048 |     0.002 |       0 |  0.770 |1219047619.0
 openssl_sha256_tl        |    2048 |     0.002 |       0 |  0.737 |1272840273.5
 openssl_sha256_noopt *   |    8192 |     0.006 |       0 |      - |1348255431.2
 openssl_sha256_opt       |    8192 |     0.006 |       0 |  0.952 |1416320885.2
 openssl_sha256_tl        |    8192 |     0.006 |       0 |  0.960 |1404423109.9
 openssl_sha256_noopt *   |   32768 |     0.023 |       0 |      - |1434173669.5
 openssl_sha256_opt       |   32768 |     0.023 |       0 |  0.990 |1449013885.2
 openssl_sha256_tl        |   32768 |     0.023 |       0 |  0.991 |1447030249.5
 openssl_sha256_noopt *   |  131072 |     0.090 |       0 |      - |1453239164.9
 openssl_sha256_opt       |  131072 |     0.090 |       0 |  0.999 |1454303371.9
 openssl_sha256_tl        |  131072 |     0.090 |       0 |  1.001 |1451726162.1
```

But that's slightly artificial, what about the minor "warm cache" (/branch predictor) win we'd expect if we digest multiple things in quick succession (simulated by "calling the function 10 times"). Then this approach is near-identical:

```
## digest sha256:

 Name (* = baseline)      |   Dim   |  Total ms |  ns/op  |Baseline| Ops/second
--------------------------|--------:|----------:|--------:|-------:|----------:
 openssl_sha256_noopt *   |      32 |     0.003 |     103 |      - |  9685230.0
 openssl_sha256_opt       |      32 |     0.001 |      42 |  0.408 | 23721275.0
 openssl_sha256_tl        |      32 |     0.001 |      44 |  0.433 | 22346368.7
 openssl_sha256_noopt *   |     128 |     0.004 |      34 |      - | 29317453.0
 openssl_sha256_opt       |     128 |     0.002 |      17 |  0.516 | 56863616.2
 openssl_sha256_tl        |     128 |     0.002 |      17 |  0.518 | 56587091.1
 openssl_sha256_noopt *   |     512 |     0.007 |      13 |      - | 75095335.9
 openssl_sha256_opt       |     512 |     0.005 |       9 |  0.721 |104086196.4
 openssl_sha256_tl        |     512 |     0.005 |       9 |  0.727 |103350827.6
 openssl_sha256_noopt *   |    2048 |     0.018 |       8 |      - |116542423.0
 openssl_sha256_opt       |    2048 |     0.015 |       7 |  0.876 |133090720.0
 openssl_sha256_tl        |    2048 |     0.016 |       7 |  0.883 |131958762.9
 openssl_sha256_noopt *   |    8192 |     0.060 |       7 |      - |136688246.7
 openssl_sha256_opt       |    8192 |     0.057 |       7 |  0.959 |142506740.9
 openssl_sha256_tl        |    8192 |     0.057 |       7 |  0.959 |142561300.3
 openssl_sha256_noopt *   |   32768 |     0.228 |       6 |      - |143992758.2
 openssl_sha256_opt       |   32768 |     0.226 |       6 |  0.992 |145198025.5
 openssl_sha256_tl        |   32768 |     0.226 |       6 |  0.994 |144806618.1
 openssl_sha256_noopt *   |  131072 |     0.900 |       6 |      - |145589128.8
 openssl_sha256_opt       |  131072 |     0.900 |       6 |  0.999 |145704360.4
 openssl_sha256_tl        |  131072 |     0.900 |       6 |  1.000 |145615331.2
```

I think that's close enough that it's worth comparing e2e numbers, so I'll try that.