@maxtropets 2025-09-05
Thanks for benchmarking, picking that up in https://github.com/microsoft/CCF/pull/7251